fixup! Make BIGNUM code available from within the FIPS module
